requirements.txt
torch
torchvision
flask
pillow
pandas
scikit-learn